Added stub of tic tac toe, and renamed sudoku solver
