+++ README.md
@@ -1,2 +1,2 @@
-# sudokusolver
+# sudoku
 Fun project to solve sudoku puzzles.
