+++ utils/utils.py
@@ -1,0 +1,3 @@
+def optionalPrint(string: str, verbose: bool=True):
+    if verbose:
+        print(string)
